Add CLOB (Central Limit Order Book) mode (#10)
* Return fills with market data

This will make it possible to display fills in the UI.

* Add get_user_active_orders()

Gets only the active orders belonging to the user

* Add CLOB simulation mode

This mode stores incoming orders in a central limit order book
and matches them off. In addition to user-placed orders, there
are also simulated orders that can match with user orders or
with each other.

* Add test_clob.py

* Add OrderSource.USER and OrderSource.SIMULATION

This will be used to differentiate between orders placed by
the user and simulated orders placed in the CLOB.